seperating grid and mesh tests
